+++ subprojects/files/src/main/java/org/gradle/api/internal/file/archive/impl/FileZipInput.java
@@ -14,9 +14,11 @@
  * limitations under the License.
  */
 
-package org.gradle.api.internal.file.archive;
+package org.gradle.api.internal.file.archive.impl;
 
 import com.google.common.collect.AbstractIterator;
+import org.gradle.api.internal.file.archive.ZipEntry;
+import org.gradle.api.internal.file.archive.ZipInput;
 import org.gradle.internal.file.FileException;
 
 import java.io.File;
